# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NORA API
  version: 0.2.0
  description: |
    API REST do backend NORA (Spring Boot 3 / Java 21). Cobre autenticacao com cookies httpOnly,
    IAM estilo AWS (grupos + policies + attachments + audit), convites por e-mail, contexto comercial
    do tenant, dominio corporativo, upload/listagem/detalhe de reunioes, gestao de tarefas extraidas,
    analise ao vivo e emissao de token efemero para Azure Speech.

    ## Autenticacao
    A maioria dos endpoints exige um JWT. Ha dois modos de transporte equivalentes:

    - **Bearer** — header `Authorization: Bearer <accessToken>` (esquema `bearerAuth`).
    - **Cookie httpOnly** — `POST /auth/login` (e `POST /iam/invites/{token}/accept`) define
      `Set-Cookie: nora_access` (`Path=/`, `SameSite=Lax`, `HttpOnly`) e `Set-Cookie: nora_refresh`
      (`Path=/auth`, `SameSite=Strict`, `HttpOnly`). O navegador reenvia `nora_access` automaticamente.
      `nora_refresh` so chega nas rotas sob `/auth` (refresh/logout). A flag `Secure` e ligada em prod
      via `nora.auth.cookie-secure`.

    O campo `accessToken` continua presente no corpo JSON do login por compat com clientes legados;
    novos clientes devem confiar nos cookies.

    ## Autorizacao IAM (ADR 0007)
    NAO existem papeis fixos. A autorizacao e baseada em **policies IAM** (estilo AWS) avaliadas por
    `(action, resource)`. Cada endpoint protegido declara na descricao a action IAM exigida
...
    `403 IAM_FORBIDDEN`.

    ## Formato de erro
...
servers:
...
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /iam/groups:
    get:
      tags: [IAM]
      summary: Lista grupos do tenant
      description: 'Action IAM: `iam:group:read`.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IamGroup' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [IAM]
      summary: Cria um grupo
      description: 'Action IAM: `iam:group:create`. Nome duplicado retorna `409 IAM_NAME_TAKEN`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateGroupRequest' }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IamGroup' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /iam/groups/{id}:
    delete:
      tags: [IAM]
      summary: Remove um grupo
      description: 'Action IAM: `iam:group:delete`.'
      parameters: [{ $ref: '#/components/parameters/GroupId' }]
      responses:
        '204': { description: Removido }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /iam/groups/{id}/members:
    get:
      tags: [IAM]
      summary: Lista IDs dos membros do grupo
      description: 'Action IAM: `iam:group:read`.'
      parameters: [{ $ref: '#/components/parameters/GroupId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { type: string, format: uuid }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /iam/groups/{id}/members/{userId}:
    post:
      tags: [IAM]
      summary: Adiciona usuario ao grupo
      description: 'Action IAM: `iam:group:add-member`.'
      parameters:
        - { $ref: '#/components/parameters/GroupId' }
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: Adicionado }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [IAM]
      summary: Remove usuario do grupo
      description: 'Action IAM: `iam:group:remove-member`.'
      parameters:
        - { $ref: '#/components/parameters/GroupId' }
        - { $ref: '#/components/parameters/UserId' }
      responses:
        '204': { description: Removido }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------------
  # IAM — policies
  # ----------------------------------------------------------------------------
  /iam/policies:
    get:
      tags: [IAM]
      summary: Lista policies do tenant
      description: 'Action IAM: `iam:policy:read`.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IamPolicy' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [IAM]
      summary: Cria uma policy
      description: |
        Action IAM: `iam:policy:create`. `document` e um documento de policy estilo AWS
        (statements com `effect`/`action`/`resource`). Documento invalido retorna `400`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePolicyRequest' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IamPolicy' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /iam/policies/{id}:
    get:
      tags: [IAM]
      summary: Detalha uma policy
      description: 'Action IAM: `iam:policy:read`.'
      parameters: [{ $ref: '#/components/parameters/PolicyId' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IamPolicy' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [IAM]
      summary: Atualiza o documento de uma policy
      description: |
        Action IAM: `iam:policy:update`. Substitui o `document` e incrementa `currentVersion`.
        Documento ausente retorna `400`.
      parameters: [{ $ref: '#/components/parameters/PolicyId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdatePolicyRequest' }
      responses:
        '200':
          description: Atualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IamPolicy' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [IAM]
      summary: Remove uma policy
      description: 'Action IAM: `iam:policy:delete`.'
      parameters: [{ $ref: '#/components/parameters/PolicyId' }]
      responses:
        '204': { description: Removida }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------------
  # IAM — attachments
  # ----------------------------------------------------------------------------
  /iam/groups/{groupId}/policies/{policyId}:
    post:
      tags: [IAM]
      summary: Anexa policy a um grupo
      description: 'Action IAM: `iam:attachment:create`.'
      parameters:
        - { $ref: '#/components/parameters/GroupIdPath' }
        - { $ref: '#/components/parameters/PolicyIdPath' }
      responses:
        '204': { description: Anexada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [IAM]
      summary: Desanexa policy de um grupo
      description: 'Action IAM: `iam:attachment:delete`.'
      parameters:
        - { $ref: '#/components/parameters/GroupIdPath' }
        - { $ref: '#/components/parameters/PolicyIdPath' }
      responses:
        '204': { description: Desanexada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /iam/users/{userId}/policies/{policyId}:
    post:
      tags: [IAM]
      summary: Anexa policy a um usuario
      description: 'Action IAM: `iam:attachment:create`.'
      parameters:
        - { $ref: '#/components/parameters/UserId' }
        - { $ref: '#/components/parameters/PolicyIdPath' }
      responses:
        '204': { description: Anexada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [IAM]
      summary: Desanexa policy de um usuario
      description: 'Action IAM: `iam:attachment:delete`.'
      parameters:
        - { $ref: '#/components/parameters/UserId' }
        - { $ref: '#/components/parameters/PolicyIdPath' }
      responses:
        '204': { description: Desanexada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------------
  # IAM — audit
  # ----------------------------------------------------------------------------
  /iam/audit:
    get:
      tags: [IAM]
      summary: Lista eventos de auditoria IAM
      description: 'Action IAM: `iam:audit:read`.'
      parameters:
        - in: query
          name: since
          required: false
          description: Filtra eventos a partir deste instante (ISO-8601).
          schema: { type: string, format: date-time }
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AuditEvent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ----------------------------------------------------------------------------
  # Invitations
  # ----------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: nora_access
      description: |
        Cookie httpOnly definido por `POST /auth/login` (e por `POST /iam/invites/{token}/accept`).
        Equivalente ao Bearer; o navegador o reenvia automaticamente. Renovado via `POST /auth/refresh`.
  parameters:
...
    GroupId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
    PolicyId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
    UserId:
      in: path
      name: userId
      required: true
      schema: { type: string, format: uuid }
    GroupIdPath:
      in: path
      name: groupId
      required: true
      schema: { type: string, format: uuid }
    PolicyIdPath:
      in: path
      name: policyId
      required: true
      schema: { type: string, format: uuid }
  responses:
    BadRequest:
      description: Requisicao invalida (`VALIDATION_FAILED`, `MALFORMED_REQUEST`, `TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Nao autenticado (`UNAUTHENTICATED`, `INVALID_CREDENTIALS`, `REFRESH_TOKEN_INVALID`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Sem permissao (`FORBIDDEN`, `IAM_FORBIDDEN`, `USER_DISABLED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Recurso nao encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflito (`EMAIL_ALREADY_TAKEN`, `IAM_NAME_TAKEN`, `INVITE_ALREADY_ACCEPTED`, ...)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, traceId, timestamp]
      properties:
        code: { type: string, example: MEETING_NOT_FOUND }
        message: { type: string }
        traceId: { type: string }
        timestamp: { type: string, format: date-time }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }

    # ---- shared enums ----
...
    IamGroup:
      type: object
      required: [id, name, createdBy, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateGroupRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
    IamPolicy:
      type: object
      required: [id, name, document, currentVersion, createdBy, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        document:
          $ref: '#/components/schemas/PolicyDocument'
        currentVersion: { type: integer }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreatePolicyRequest:
      type: object
      required: [name, document]
      properties:
        name: { type: string }
        description: { type: string }
        document: { $ref: '#/components/schemas/PolicyDocument' }
    UpdatePolicyRequest:
      type: object
      required: [document]
      properties:
        document: { $ref: '#/components/schemas/PolicyDocument' }
    PolicyDocument:
      type: object
      description: |
        Documento de policy estilo AWS. Lista de statements avaliados por `(action, resource)`.
        Aceito/retornado como JSON livre; a forma canonica abaixo e ilustrativa.
      properties:
        version: { type: string, example: '2025-01-01' }
        statements:
          type: array
          items:
            type: object
            properties:
              effect: { type: string, enum: [ALLOW, DENY] }
              action:
                type: array
                items: { type: string, example: 'meeting:read' }
              resource:
                type: array
                items: { type: string, example: 'nora:tenant/{tenantId}:meeting/*' }
      additionalProperties: true
    AuditEvent:
      type: object
      required: [id, actorUserId, action, createdAt]
      properties:
        id: { type: string, format: uuid }
        actorUserId: { type: string, format: uuid }
        action: { type: string, example: 'iam:policy:create' }
        targetType: { type: [string, 'null'], example: POLICY }
        targetId: { type: [string, 'null'], format: uuid }
        payload:
          type: [object, 'null']
          additionalProperties: true
        createdAt: { type: string, format: date-time }

    # ---- invitations ----
...